chore: separate solutions, and tests into their own files, refactor, and update readme with links
