+++ java/README.md
@@ -30,4 +30,21 @@ mvn test
 - [345. Reverse Vowels of a String](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/ReverseVowelsOfAString.java)
 - [2299. Strong Password Checker II](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/StrongPasswordCheckerII.java)
 - [653. Two Sum IV - Input is a BST](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/TwoSumIV.java)
-- [...](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/***.java)
+- [543. Diameter of Binary Tree](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/DiameterOfBinaryTree.java)
+- [226. Invert Binary Tree](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/InvertBinaryTree.java)
+- [2583. Kth Largest Sum in a Binary Tree](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/KthLargestSumInABinaryTree.java)
+- [107. Binary Tree Level Order Traversal II](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/BinaryTreeLevelOrderTraversalII.java)
+- [102. Binary Tree Level Order Traversal](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/BinaryTreeLevelOrderTraversal.java)
+- [145. Binary Tree Postorder Traversal](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/BinaryTreePostorderTraversal.java)
+- [94. Binary Tree Inorder Traversal](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/BinaryTreeInorderTraversal.java)
+- [144. Binary Tree Preorder Traversal](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/BinaryTreePreorderTraversal.java)
+- [67. Add Binary](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/AddBinary.java)
+- [290. Word Pattern](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/WordPattern.java)
+- [520. Detect Capital](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/DetectCapital.java)
+- [14. Longest Common Prefix](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/LongestCommonPrefix.java)
+- [412. Fizz Buzz](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/FizzBuzz.java)
+- [709. To Lower Case](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/ToLowerCase.java)
+- [28. Find the Index of the First Occurrence in a String](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/FindTheIndexOfTheFirstOccurrenceInAString.java)
+- [171. Excel Sheet Column Number](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/ExcelSheetColumnNumber.java)
+- [2243. Calculate Digit Sum of a String](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/***.java)
+- [...](https://github.com/jizzel/algo-dsa/blob/main/java/src/main/java/org/algodsa/)
