+++ java/src/main/java/org/algodsa/solution.java
@@ -1,0 +1,17 @@
+package main.java.org.algodsa;
+
+/**
+ * Leetcode Java Solutions
+ */
+public class solution {
+
+    /**
+     * Default constructor for the Solution class.
+     * This constructor initializes the Solution class.
+     */
+    public solution() {
+        // Initialization code (if any) goes here
+    }
+
+
+}
